base: vector: add function fromPolar as an extension of fromAngle
And use it everywhere where it is appropriate.
This is a very common operation and was previously done in three
different ways:
* Vector.fromAngle(angle) * length
* Vector.fromAngle(angle).scaleLength(length)
* Vector.fromAngle(angle).withLength(length) (previously setLength)

The last one is inefficient, but it is also confusing to use different
ways to achieve the same goal. fromAngle should from now on only used
sparingly, where necessary, while most code should use fromPolar