base: make vectors readonly
There are no functions to modify a vector anymore.
The few cases of manual setting of the x and y coordinates are adapted
in this commit, but quite few were actually present.

This makes vectors easier and most importantly safer to use, avoiding
weird side effects on the mutation of vectors