amun: trajectorypath/speedprofile: refactor slow-down calculation to a single class
Previously, the computation was scattered among the different functions
timeWithSlowDown, endOffsetSlowDown, positionAndSpeedForTimeSlowDown
etc.

By making these functions generic and supplying classes for the actual
acceleration computations, the code is much more readable and
maintainable (and also shorter).
The performance is not impacted.

Note that this is not an equivalent transformation since several bugs in
the old code are fixed by this more generic approach. But is maintains
the expected correct behavior of the functions.