amun: trajectorypath: remove indicator value (0, 0) for desired distance
It could possibly lead to incorrect behavior when the task is to drive
to the current robot location.
While the desired distance was previously not used when constructing a
trajectory from time and angle, it is ok to do so since the actual
distance of that trajectory can be used (it is usually computed
anyways).
With that, the trajectory will not be changed.