amun: trajectorypath: add start time to trajectory input
Missing this was a bug since samplers like the standard sampler can run
"after" the endinobstacle sampler, after it found a trajectory out of
the obstacle. In these cases, the StandardSampler still checked the
moving obstacles without a time offset.

(Also do not store TrajectoryInput on the Trajectorypath class, this is
not necessary)