Ra: logopener: move showLastPosition
Previously, the logopener used the statusSource itself to determine if the
lastPosition button should be shown or not.
As the statusSource will be living in amun soon,
we change to use the information that is already published to the logslider
to be able to calculate its labels and limits.

The minor change in logslider is due to the fact that the current implementation
did not forward this status packet, as it wasn't used outside the slider.
This will be changed naturally when all Status come from seshat,
as the forwarding part in LogSlider (and its TimedStatusSource)
are going to be removed.