Ra: logopener: replace m_logFile by a boolean
m_logFile has not been used in logopener other than to check if the weak_ptr still holds a reference.
This function boils down to logslider still holding the reference to that statusSource in its timedStatusSource.
There are two reasons why the statusSource could go out of scope:
A) we opened another file or
B) we used instant replay.

In both of these cases (and when destructing the logopener) we want to write the current position aswell.
So I removed watching over the reference in favor of setting and resetting a boolean in
openFile(QString) and saveCurrentPosition respectivly.