amun: tracker: use own lastUpdateTime for each camera
Otherwise, packets will be discarded in real life due to network
latencies and similar issues. It can also be a problem in the simulator,
if the packets for both cameras are created with the exact same time
(can happen depending on the system timer accuracy).
Before, with a less precise timer, it could happen that all packets from
cameras except zero were discarded since they all had the same time.