glados: a/attacker: Move FreeKick and DoubleTouchGuard further down
In the future, both the `FreeKick` and the `DoubleTouchGuard` behavior
will be moved from the normal behavior list to an objective. Since
objectives are run at the bottom of the behavior list, both need to be
moved further down.

Because of this, behaviors that were previously below of `FreeKick` and
`DoubleTouchGuard` need to make sure they don't run during a freekick.
- `BallEscort` and `Duel` need a check
- `DuelAssistant` is never selected by the main attacker, `FreeKick` is
  only selected by the main attacker. It only needs a check for double
  touch, since the double touching robot does not have to be the main
  attacker.