ra: teamwidget: make default text shorter
When opening ra, the default text (Strategy Disabled and <no
entrypoint>) will be shown for a very short time before the strategy is
loaded. This resizes the widget and makes it impossible to permanently
have the widgets at a small size even though the strategy name and
entrypoint would allow it.