glados: tr/trainer: dump inbox only once
The second dump inbox was kept since the trainer can send messages to
itself which would not be visible otherwise. Since these messages are
now dumped directly by the messaging, it is no longer necessary to dump
all of these messages twice. This reduces log size and improves the
run-time.