ra/fieldwidget: option to show vision data
"Tracking From" can now be set to None and a checkbox controls the
visibility of the vision objects.
Refactored creation of robot visualization to a separate function for
use in setVisionRobot. The new bool changes whether the id is shown and
sets the transparency accordingly.
The graphics objects of the vision balls and robots are stored in a map
that uses the camera id as a key to avoid flickering between camera
frames.