amun: s/seshat: add checkForwardImportance
This function has to emulate the behavior currently implemented
by forwarding amun::gotReplayStatus directly to the
replayteamwidgets.

It contains the same bugs as the original solution,
but including a new comment that describes them.

This commit is needed as gotReplayStatus is bound to disappear
as the whole replay-logic moves purely into Amun