glados: task: Construct with behavior instead of agent and don't cache messaging
The messaging object was (and, at the moment, is) often cached on tasks.
This prohibits behaviors from changing their agent since the messaging
object depends on the agent and the cached version would belong to an
old agent instance.

The robot accessor was deleted since the robot should rather be accessed
over behavior().