glados: a/b/behavior: Allow changing the agent of checkables and don't cache messaging
Not caching the messaging object is not strictly necessary. The behavior
could overwrite the cached reference in `setAgent`. I did however change
this for consistency with 60b9a40a79 and b8b0fa0064.

The robot() accessor on Behavior was deleted since the robot should
rather be accessed over agent().robot() and was not used anyway.