glados: g/centerback: use robotMinEndspeed
Previously, the trainer would abort locking the centerbacks because " the robot is not going to make it".
However, the timing calculations differ in g/centerback and t/d/centerback, as latter will use
robotMinEndspeed to improve their speed if the need arises. As g/centerback assumes
the centerback has to slow down to 0, its timeAroundDefenseArea will increase even further.

While is does not really make sense if there's at least one corner between the robot and cBPos,
it improves the cases where the ball is already very close and while the robot isn't perfectly centeres,
it will hit the ball quite reliably.

Not locking the robot can result in moving the robot to a completly different location,
and while the ball still will hit the robot, it will stop sending a shoot command
because we're now " far away from our destination "

If this patch fails, it might be an option to lock necessary / unnessesary
like done in a/a/shoot, or modell the robot as an physical object (instaed of the current point - model)
to make sure the necessasy is not dropped if the ball will hit the robot at the dribbler,
even if the ball will not be centered