glados: a/s/breakpass: Don't break passes as main attacker
This makes no difference, since both attackers and defenders have
behaviors (respectively a/a/shoot and a/d/handleball) acting as default
behaviors for main attackers which are higher up the priority list.
However, since I want to move a/a/shoot into an objective, and since
a/a/runobjective is lower on the list, this needs to be explictly
checked.