Glados: g/moves: always update _numAttackersSent
This was a simple oversight when this bool was introduced in bc29b6670.
If a move is aborted because one of its members ceases to exist,
we choose a new move. To make sure not to start the move eagerly,
while defenders are beeing recruited to be part of that move in the next frame,
we have to reset _numAttackersSent