Glados: te/u/g/pools: don't use zombie-agents
The old version of this test uses a zombie-agent
to send the second pool-change-request:
After the first pool change has happend,
the coordinator created two new agents
(one for each robot that changed the pool).
When it is time for the second pool change,
these old agents are no longer handled by the coordinator,
and it's quite unclear if failing to do so is an issue
in a normal game of SSL.

Instead, we get the agent for a robot after each pool-change.
This also allows us to test the correct insertion in each pool more
frequently.