Glados: c/maincoordinator: handle swapping robots
Previously, the code in maincoordinator was not able
to handle swapping robots, i.e. where one robot wants to become
a defender and another robot wants to become an attacker (see testcase
introduced in ef582e3809a04074cea9abd8b491e14a4ecacabc).

While the previous implementation could just work fine if these swaps
would be a possibility, it was up to luck whether both changes went through fine,
or whether the first change dropped the robot that was about to be chaning pool
on its own command. If that were to happen, the robot would not be in its old pool
(as it was already decided that the robot has to change), so "invalid pool change request"
would fire.