glados: c/agentpool: Don't drop attackers participating in a move
This is the completion to 5e3bca7e7: While 5e3bca7e7 did allow moves to specify ALLOW_EXTRA_ATTACKERS,
this never really worked in all cases, becase the attacker participating in this move (at the moment, the only
move that specifies that is ballplacement), might be the \"worst attacker\", so if attackratio > move.N_ATTACKERS,
and attackratio changed from lets say 5 to 4, one of the particiaption robots might be demoted to defender,
and therefore the move would reset.

It was considered, but disregarded, to use forceKeepingInPool on all agents that participate in a move.
This is mostly due to the fact that forceKeepingInPool always has one frame delay and is therefore used
to control which robot should get demoted when the move ends and its forced number of attackers drops
(this is mostly the case for moves executed as a set-play: Kickoff or Direct).

Instead the message "moveInfo" now contains the exact robots that are participating in the move
(compared to before, where it just contained the number of attackers).
In the first frame of the move, the size of the array indicates how many
robots are asked for, and each position in the array is filled with
undefined. In all other cases, the ids of the participating robots are
included.

Using that message "moveInfo", that now has to be sent to all robots
instead of only sending it to the trainer (see comment in messaging for
details), the new subclass AttackerPool can decide if a robot is not up
for debate, but has to stay in this pool (which happens if the robot
participates in a move).