amun/trajectorypath: escape obstacles: mark trajectory as safe if stops outside the obstacle
Before, it was necessary to have the trajectory leave the obstacle for
at least 1.5 seconds. If not enough space is provided for a 1.5 second
long trajectory segment, no safe trajectory was found and the escape
obstacle sampler failed, even if some checked trajectories actually
ended outside the obstacle.

By making all trajectories safe that end outside the obstacle, this
problem is avoided.