glados: tr/roles: Assert that a robot doesn't apply twice
The trainer would just use the last sent application. This is probably
an error since it relies on the fact that collections are processed in
insertion order and this not obvious here.

Note that if this behavior is wanted in the future, its probably ok to
remove the assertion. I changed this not because of a found bug, but
because since d6502f5a18 this would no longer be visible in the debug
tree.