trajectorypath/base: add radius to rectangle obstacles
The radius was already in use internally in the obstacle, it was just
not possible to construct one with a radius != 0.
To make it possible to use this radius for productive purposes in the
strategy and make it clear that the radius has to be considered when
writing functions for the rectangle obstacles, it is not necessary to
set this radius.