trajectorycli: add optimizer for the alpha time trajectory generation parameters
The parameters there were already optimized previously, but the
optimizer code was never published as it required inconvenient code
changes in several places. With the trajectorycli framework as it is,
the optimizer can now be used properly.

The optimizer code was generalized for both the end in obstacle
optimizer and the alpha time trajectory optimizer.

Also the printed paramters are now sorted, to make it easier to
correctly write the optimized parameters into the code without errors.