ra/amun: implement tracking replay
The tracking replay enables users to re-run the tracking on a given log
file without losing information such as the exact execution time of the
tracking. This is not possible when converting to vision logs and
executing the tracking on them.

The current state is not complete however, as only the ball is currently
perfectly re-played, while the robots still have a small deviation
between the original log positions and the re-played log positions.
The reason for this is not yet known and the issue can therefore not yet
be closed.