amun: save and use radio command set times for tracking replay
These timings were previously not accessible and a perfect tracking
replay for the robots was thus not possible.

Another change here is that status_debug and status got merged in
Processor::process.
This is important as the radio commands used for the tracked state in
status were in statud_debug.
If status_debug is emitted only after status, the radio commands used
for the tracking replay will always be off by one frame.