Simulator: Don't try to send commands to non-existing robots
This is a pure precautionary measure: While I neer have seen this,
there might be a race-condition, where the simulator already destroyed
a SimRobot object, but still recieves an old robotCommand, sent by a strategy
that still had this robot. The old implementation might have segfaulted,
an unaccaptable result for ra.

While I don't like the silent discarding of commands, this is more or less identical,
to what happend on the old implementation before
73ee8d549dd70b617be5a24bed723dccc47256d9.