seshat: combinedlogwriter: do not save status packages with time 0
If the logging was enabled before the first status package arrived, the
team status package created by sendIsLogging(true) used m_lastTime,
which was zero, thus saving a status package with time zero in the
logfile.
This is problematic as it effectively corrupts the log file (or at least
the first block in the log file).