seshat: logfilereader: report logs with packets with timestamp 0 as invalid
They were already invalid before in the sense that the block in the log
containing the packet with timestamp zero was ignored (empty status
smart pointers were returned). Now there will be an error message
instead, with the opportunity to repair the log by hand if necessary
instead of just having some status messages disappear.