tracking: ballflyfilter: set an estimate of the chip start time
... before accepting the chip.
This is necessary since detectionPinv(...) uses this variable for its
decision although the time is only properly set after the chip is
accepted.
Without this change, the time would be random for the first chip and the
start of the previous chips in the next ones, not the intended behavior
of the variable.