glados: group/move: More attackers for offensive kickoff
New method wantedMaxRobots in base - when not implemented returning
MAX_ROBOTS, otherwise the amount of wanted robots based on accessible
robots

In kickoff a flexible amount of 2 to max 5 attackers are assigned a
special position for kickoff. By wantedMaxRobots the number of attackers
is decided to ensure a solid amount of defenders as they are assumed to
be more important than additional attackers. Positions are stored in
_assistantAndPasPos together with the required pass position and are
assigned iteratively to all acquired robots. Fixes #786