glados: a/s/ballescort: Send an objective as defender
The main attacker should always send an objective. For attackers, this
is performed by `a/a/selectobjective`. For defenders, main attacker
behaviors need to be adjusted. If the main attacker is defender, he may
take one of two behaviors. `a/d/handleball` is the default case and was
addressed in 838758a0ee. This commit addresses ball escort.

Neither Striker nor Midfield are particulary sensible choices but
Midfield is probably a good default objective to send.