glados: t/a/volley: Move to u/volley and remove _volley function
Every function in `t/a/volley` was static except `_volley` and that was
unused. Thus, I transformed it into a module with free functions

The code should be functionally equivalent, but the diff is large since
I had to delete one level of indentation.

In the past, `task/shootgoal` decided whether to shoot a volley and
calculated relevant positions. This was buggy (see #250) and
subsequently deleted (see 7a6e0443). After that, the `_volley` function
was not in use anymore. Instead `CatchBall` would use `calcPhi` to
calculate the direction the robot should face (ignoring the returned
shoot velocity). `t/a/shoot` would then use CatchBall and shoot the ball
when nearby.

Volley handling in `t/a/shoot` was later improved in 8534a894