base: robot: Add an interface for physical robot state
The intuitive usage for `moveObjects` would be to
- first modify the Robot objects in `World.FriendlyRobots`
- then pass these instances to `moveObjects`
but modifying these global objects is not a good idea in most cases.

Thus, we provide a function to copy the relevant state of the Robot
instance, modify the copy and pass that to `moveObjects`

More state may be added to `RobotState`, however that is not needed
right now.