docker: Specify correct build context in build hook
It is required that some images (in this case: simulatorcli) are built
from the repository root. Well, how do you specify that in the Docker
Hub Automated Build settings? Oh, you can specify a build context in the
settings and a Dockerfile path relative to that. How nice!

But, alas, we need BuildKit to be enabled. Why isn't that the default
for Docker Hub builds you ask? I don't know. Maybe it is, but that is
not documented.

Custom build hooks to the rescue. But the build hook fails. Why? "No
such file or directory: data". Whats the working directory of the build
hook? Oh, thats not documented. Maybe it is the build context specified
in the build settings? At least the option says *build context*, the
Dockerfile path is relative to that, and the $DOCKERFILE_PATH
environment variable available to the hook is that relative path. No,
it's not the build context. Maybe it's the hooks directory, at least
that is the location of the script? No, not that either. Printing the
working directory inside the build hook did not work half the time.
Obviously, the working directory is the location of the Dockerfile.