amun: p/processor: Detect division and amend status accordingly
Since the division is not part of the referee message the division is
inferred via the field dimensions. This information is then added to the
geometry message as an additional field to avoid computing it again at
every point that needs it.
The "Team Information" widget is now the "Referee Information" widget,
because it now also displays the current division, which is obviously
team independent.

strategy/base/world: add support to read the division from the geometry
message.