amun/ra: handle yellow cards only in the GC
There is no need to have the yellow card logic in the UI anymore, since
it is now done with the game controller. If the GC is disabled, the card
times will not count down.

The change in the teaminfowidget is fixing the bug that the current
cards label will never go down, even when a card times out, since the
card counter in the SSL_Referee packet is the total number of cards
issued in the game and not currently active number.