docker: simulator-cli: Use tini as dedicated init process
Without this fix, `docker stop` would not work on a container running the
image.

`docker stop` tries to stop the container by sending SIGTERM to the
containers init/PID 1 process. However, kill(2) states

> The only signals that can be sent to process ID 1, the init process,
> are those for which init has explicitly installed signal handlers.
> This is done to assure the system is not brought down accidentally.

Since the image's entrypoint is used as init process in the container,
it automatically ignored SIGTERM.

Thats also the reason why `simulatorcli_entrypoint.bash` does not exec,
but instead forks. While testing I noticed that, when the bash script
would exec, pressing Ctrl-C (aka sending SIGINT to the foreground
process and all it's descendants) would not terminate the container.
The PID 1 process (bash) would ignore the SIGINT. The simulator-cli
received it, causing it to terminate and thus causing the bash script to
continue execution and terminate itself.

That is of course, only half the fix since SIGTERM still wasn't handled.

You could
- exec in the bash script, making the simulator-cli the init process and
  handle signals there
- Patch bash to relay SIGTERM

Both are bad ideas

- Handling signals correctly is hard to do in a QT application. All QT
  functions (including `QCoreApplication::exit(int)`) are async signal
  unsafe [1]. A possibility would be to write to a Unix Pipe and handle
  writes on the other side of the pipe in the QT event loop (e.g. using
  QSocketNotifier), however, that seems no portable. I think the best
  bet would be to rely on QT handling the signals, or at least to rely
  on it giving a portable way to relay signals to the event loop, but
  this is not currently the case [2].

- Writing a correct custom init process is hard (has to handle child
  proceses correctly)

We thus use tini as dedicated init process. It will simply relay signals
to child proceses and correctly handle SIGCHLD. Note that the problem
described in [2] is still there.

[1] https://doc.qt.io/qt-5/unix-signals.html
[2] https://bugreports.qt.io/browse/QTBUG-10928