ra: refereewidget: Fix a situation where setting the divison did not work
Refereewidget does not only change its division text when
manually selected, but also when the division changes, because of
somewhere else to make it consistent with the actual current division,
even if none was set by the user.
The bug was that the division changed back if there where still packages
processed in other threads, because setCurrentText in handleStatus also
triggered the command to change the division.
Now a bool differentiates if the signal comes from user input or from
handleStatus and ignores the latter.