amun: Change default port where vision detections are received
According to https://ssl.robocup.org/league-software/ vision detections
are sent on port 10006.

This previously worked probably because Ra has 10006 as the default
vision port (at least in ra/configdialog.cpp, I don't know what the
value in configdialog.ui is used for) and it sends it's config once on
startup.