glados:group:move:hardwarechallenges: add base class for hardware challenges
The HardwareChallengeBase class takes a JSON for the constructor and
provides a mechanism to fullfill the requirements of this JSON.
If it is run in the simulator with debug mode on the robots get
teleported to the correct positions.
Otherwise (since we won't have enoug moving robots in time) one
friendly robot moves to a position that allows placing the static
obstacle accurately and when the obstacle is placed moves on to
the next and so on.
Afterwards the friendly robot moves to its starting position if one was
set in the JSON.
The state is then initialized and challengeSpecificUpdateTask is called.