amun: typescript: extend functionality of the require function
This allows for a list of files to be required in the same context.
Using this is of course only reasonable when cleanRequire is set,
otherwise, it has the same effect as just using multiple requires.
This change allows for requiring multiple files in the defined order in
the SAME clean require context.