ra: fix ball placement by using the fieldwidget flipping the field
This was because the fieldwidget sent a referee packet that only
contains the ball placement information, while the internal game
controller needs the referee information to be accumulated (like the
internal referee in the ui does). Therefore, the ball placement
information is now routed through the internal referee and merged with
the accumulated referee info.