amun: simulator: fix error in calculateDribblerMove
The constructor btQuaternion(btVector3, btScalar) is not for imaginary
and real part respectively, but instead expects a rotation in axis-angle
representation.
getAxis() follows the same logic and does not return the imaginary ijk
vector, but the axis of the axis-angle rotation.
Additionally rotating a vector p with a quaternion q is q*p*q^-1 and not
q*p.