glados: c/agentpool: fix crash when robot tries to become attacker while move is running
The situation was the following:
* a move is currently running and using 5 robots
* a defender wants to become attacker and sends a poolChangeRequest
* hasExchangeableRobot returns true because although the robots from the
move are set as immune to pool removal, they do not set
forceKeepingInPool
-> the coordinator now tries to remove one of the move robots to
accomodate the new one, which results in the crash: Error: More immunes
than allowed Robots: 4

The solution is therefore to check for immunity in hasExchangeableRobot.

This commit also removes the check for amin.isDebug in the exchangeRobot
function, since not using this way out of the function leads to more
crashes, which is the wrong way around.