glados: c/agentpool: fix crash with negative robot limit
Fixes #791.

At the time of calling hasExchangeableRobot, the agent pool can still
contain more robots than the pool limit, which can result in a wrong
return value.

The crash "We do not allow negative robot limits: -1" was caused when
the pool limit was zero and there was still one robot in the pool that
could be removed or exchanged from the pool (which it would habe been
later on in the frame anyways).
Since another robot wanted to change into this pool at the same time,
maincoordinator:_changeRobot got to
this._pools[newPool].setRobotLimit(poolLimit - 1);
with poolLimit - 1 being -1, causing the crash.