amun: fix replay not working after switching logs
If the first log that played was a log with replay information (the
corresponding strategy was running before), the strategy switched to
only using the status packages that contain this information.
If the second log was now a log without them (for example a converted
vision log), then the replay strategy would ignore everything, waiting
for the replay information.

Therefore, the isReplay value in the strategy must be reset when
switching between status sources for playback.