glados: o/robot: improve own standard shooter with invisible ball
If the robot obscures the ball while driving close to it, hasBall never
activates and the standard shooter is not identified.
Given that we do not drive to the ball with more than one robot, this
rather loose estimate is fine as well.

Without this, we tend to produce more double touches.