glados: a/b/agent: More often allow MA applications if ball is in our defense area
Most of the time, if the ball is in our defense area, we want the keeper
to be the main attacker. This is not the case while we are in a stop
state. (making the keeper main attacker would only be sensible if we
intend him to drive to the ball somehow; however, the only one "driving
to the ball" in stop states is the stop attacker)

Fixes #835
The ball was occluded by an opponent robot and thus assumed to be in our
defense area. Because of this, the sole attacker did not become main
attacker and no objective was sent. Because of this, the attacker had no
behavior which lead him to commit ball placement interference.