base: field: invert meaning of the given extra distance/radius in limitToAllowedField
Previously, a positive number would move the field boundary further
inwards.

This change is made to make the function consistent with limitToField,
isInAllowedField and others. For these functions, the value works the
other way round, a positive number moving the field boundary outwards.

The change is important, as exchanging limitToField with
limitToAllowedField might be done without consulting the exact
documentation of the functions, which would have been dangerous before.