glados: t/a/shoot: fix shooting ball from outside of the field
There were two problems in the code.
The first is that the extra distance in limitToAllowedField was the
wrong way around, limiting the position to something slightly outside of
the field.
The second issue is that limitToAllowedField does not limit along the
ball speed, and the resulting position is therefore not on the future
ball line.
When the ball travel time is calculated to compute the future ball,
there will therefore be some additional distance that can put the future
ball outside of the field even if the limited position is inside the
field.