ra: mainwindow: disable simulator config widget at startup
The enable simulator action is not active at startup, and the connected
signal activates only when it is toggled (when during loading, the
action is enabled). The default state of the action and the widget must
therefore be the same.

The signal/slot connection between the action and the widget is created
in the mainwindow.ui file and not in the cpp code.