amun:tracking: change reset behaviour to fix bug on teleport
Previously teleporting a robot could lead to weird orientations, because
the tracker was not reset correctly. This caused the tracker to for
example get one vision package before the command and then one
afterwards and sometimes determined that the angular velocity between
those two frames is to large, which caused a radio command to 'correct'
this.
Now the reset happens only after the first vision package that is new
than the reset command.
m_resetTime was changed to m_timeSinceLastReset to avoid confusion with
the new variable m_timeToReset.