glados: t/a/catchball: fix driving to wrong destination with intermittent activation
The problem was the following:
* Catchball was active for some time while the ball was still quite far
away from the robot (and thus the last reasonable ball pos was always
updated the actual ball pos)
* t/a/shoot decided to do something else and did not activate catchball
for some time (since it was not active, the last reasonable ball pos did
not change and stayed at the last position, quite far away from the
robot and now also from the actual ball)
* catchball gets activated again, but the ball is already "in" the
robot, meaning that the last reasonable ball pos did not get updated
anymore and stayed at the outdated position far away from the robot.
This position now got used as the ball position to drive to since the
ball is "inside" the robot, which obviously results in bumping against
the ball in the wrong way.

The solution is to always update the last reasonable ball pos,
regardless of whether catchball is running or not.
Since the specific robot for catchball is then not known, just run it
for all of them. This also allows for some simplification of the
function.