glados:hwchallenge: recheck opponent pos and move friendlies sooner
The opponent positions were previously only checked once on placement,
but in case the robot or a human accidentally moves the obstacle this
was not enough. Now the positions of all obstacles get checked every
frame until everything is in place.
The friendly robots now move to their designated positions instantly
instead of waiting for the obstacles and ball. This moves them out of
the way to make placement easier.