glados: a/b/agent: make exit condition for ball in defense area the same as a/k/handleball
This makes it so that the other robots are allowed to be MA exactly when
the keeper is not interested in being one.

This fixes the robots executing a/a/default for too long, since when the
ball was in the defense are and fast, no main attacker was selected (the
keeper did not want to be MA because the ball was too fast and the other
robots did not want it because it was in the defense area).