amun: tracking replay: reset tracking when going back in the log
Before, it was not possible to go back in the log while using tracking
replay since the tracking discards all packets that are older than the
newest one it has received.
By resetting the tracking when going back, this problem is fixed,
although the state after going back might not be the same as originally.