strategy: distinguish between internal and external simulation
This allows a strategy to handle an external simulation
like a real field for some applications,
and like the internal simulator for other applications.

The most visible application where a strategy might be interested
in treating the external simulation like a real field
is moving your robots in the UI while no control-connection
with the external simulation has been established.