amun/seshat: fix buffering of strategy status updates in horus mode
They need to be always sent when switching back from horus to ra, as
these are the status messages that contain potential compilation updates
from the strategies while horus mode was enabled.
Previously, they were only sent when seshat forced ra mode (for
switching back to ra during tournaments when they continue). This was a
bug introduced in the seshat extraction from the ui.