glados: a/b/agent: applyformainattacker: limit rating boost
This is to prevent the rating boost from overshadowing the entire
timeToPos calculation. With how it was before, if the ball was at 3 m/s
for example, the rating boost could range from -1.5 to 1.5 depending on
the orientation of the ball and robot.
Consider the case if two robots that apply for being MA stand directly
on the line of the robot speed, but one is very close and the other is a
few meters away from the ball.
Since the rating boost is 1.5 for both robot, the timeToBall is reduced
to 0 and both get an equal rating of 1.
This makes the selection between the two robots effectively random, or
rather is prefers the previous MA if it was for example the robot
farther from the ball.
This is even though the robot close to the ball would be the ideal MA
candidate no matter what the previous MA was.
An example for this situation can be found in the mainattacker unit
tests, added in the previous commit.