glados: a/a/selectobjective: use the same ballposition for canStart and canContinue
This commit fixes two aspects of the one issue:
The running objective was midfield (which uses the given ballposition to
check if it can start).
Since lastIncomingPassInfo was not executed every frame, but only if the
robot was the mainattacker, the last incoming passinfo was outdated and
far in the own half of the field (a position for which midfield happily
activates). However the ball itself was far in the opponent field half,
so far that midfield would return a canContinue of false.
This lead to the problem of the midfield objective being reconstructed
every frame, since canStart always returned true, but canContinue always
returned false.
This inconsistency can be fixed using the same ball position for both
functions.

The second issue was the wrong value of last incoming pass info.
This is solved by always calling this function at the start of
selectObjective, so that it does not miss out on updates in frames where
the robot is not yet the mainattacker.