amun: tracking/ballcollisionfilter: Improve dribbling capabilities
By checking possible ball visibility of the initial ball position where
the ball collided with the robot (or was pushed to by the robot).
It is assumed that a robot is not dribbling with the ball until this
position is shown to not have the ball once the position is no longer
occluded by the robot.