glados: t/a/stopattack: stop flickering when opponent wants to pass staight up
The issue there was that 2PI is just next to 0,
but bound(min, 0, max) is min and bound(min, 2PI, max) is max.

Instead, using the new funtion angleBound on geom,
which understands what angles are and how to bound them,
makes sure that bound(min, 0, max) === bound(min, 2PI, max).