glados: o/ball: isShot: add condition that the ball must not be too close to the robot
This is to prevent isShot from activating from smaller tracking speed
deviations that can occur after stopping the ball and it jiggling a bit.
hasBall uses .3 seconds, that must be enough to travel the required 8cm
if the ball is actually shot.

Also clean up the code a bit, some of the variables where only saved for
the debug output, which is mostly unnecessary and can thus be removed
(after all, this function can be perfectly replayed).