glados: o/ball: Check whether dribbler edges are in the receivesPass cone
Especially when the robot is close, the dribbler and robot center may
not be in the cone, even if the ball will hit the robot. This leads to
false negatives. Also checking the dribbler edges should circumvent
this. It may lead to more false positives, however that is preferable to
the contrary.

This diff also reformats the comment