glados: a/a/stop: continue running if ball is at freekick pos again
Consider the following scenario:
The opponent has a freekick, they have not yet driven to the ball.
There is a ball-misdetection in the oponnent robot dribbler and the ball
"moves" more than 5 (10 in the cpp code) cm as told by the tracking.
amun/processor/referee then switches the game state to game and we will
drive to the ball.
Therefore, consider the freekick still ongoing if the ball is at the
position as at the start of the freekick again (with some conditions of
course).