ra: plotter: add plots for simple tracking of robots
If we want to know what the robots actually do, the speed shown by the
robot tracked with radio commands can be very unreliable when the robots
do not perform quite es well as expected by the accelerator. Therefor,
it is good to also have the robot (speed) tracked without the radio
commands.
While this information is also partially available from the radio
commands, that is harder to use and parse.